Corregí el problema de la suma realizada en SorpresaDdePuntaje, agregando un casteo como float. Adicionalmente, ordené un poco la clase Jugador. Gonzalo 23NOV2013

+++ trunk/src/algoritmos/gpschallenge/modelo/sorpresa/SorpresaDePuntaje.java
@@ -7,6 +7,6 @@ public abstract class SorpresaDePuntaje extends Sorpresa {
 	protected int mPorcentajePenalizado;
 	
 	public void aplicarA(Vehiculo vehiculo) {
-		vehiculo.addMovimientosBonus(vehiculo.getMovimientos() * this.mPorcentajePenalizado / 100);		
+		vehiculo.addMovimientosBonus((float)vehiculo.getMovimientos() * this.mPorcentajePenalizado / 100);		
 	}
 }
